pyproject.toml
[tool.poetry]
name = "cc-jupyter-service"
version = "0.1.0"
description = "CC-Jupyter-Service is part of the Curious Containers project. It implements a web service to execute jupyter notebooks on a cc-agency installation."
authors = ["Christoph Jansen <Christoph.Jansen@htw-berlin.de>", "Bruno Schilling <Bruno.Schilling@student.htw-berlin.de>"]
license = "AGPL-3.0"
readme = "README.md"
repository = "https://github.com/curious-containers/cc-jupyter-service"
homepage = "https://www.curious-containers.cc/"

[tool.poetry.dependencies]
python = "^3.5"
flask = "^1.1.1"
nbformat = "^5.0.4"
jsonschema = "^3.2.0"
"ruamel.yaml" = "^0.16.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
